EPMRPP-49362: Make IntegrationBreadcrumbs a common component and expose to plugins
