+++ app/src/components/main/simpleBreadcrumbs/simpleBreadcrumb/simpleBreadcrumb.scss
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-.integration-breadcrumb {
+.simple-breadcrumb {
   display: flex;
   align-items: center;
   padding-left: 5px;
