+++ app/src/pages/common/settingsPage/integrationsTab/integrationsTab.jsx
@@ -27,8 +27,8 @@ import {
 import { showDefaultErrorNotification } from 'controllers/notification';
 import { availableGroupedPluginsSelector } from 'controllers/plugins';
 import { PLUGIN_NAME_TITLES } from 'components/integrations';
+import { SimpleBreadcrumbs } from 'components/main/simpleBreadcrumbs';
 import { IntegrationsList } from './integrationsList';
-import { IntegrationBreadcrumbs } from './integrationBreadcrumbs';
 import {
   INTEGRATION_SUBPAGE,
   INTEGRATION_SETTINGS_SUBPAGE,
@@ -131,10 +131,7 @@ export class IntegrationsTab extends Component {
     return (
       <div className={cx('integrations-tab')}>
         {subPage.type && (
-          <IntegrationBreadcrumbs
-            descriptors={this.getBreadcrumbs()}
-            onClickItem={this.changeSubPage}
-          />
+          <SimpleBreadcrumbs descriptors={this.getBreadcrumbs()} onClickItem={this.changeSubPage} />
         )}
         {this.getPageContent()}
       </div>
